requirements.txt
requests
beautifulsoup4
pandas
matplotlib
selenium
pyserial
numpy